requirements.txt
pandas==2.1.4
matplotlib==3.8.2
pandas-datareader==0.10.0
python-dotenv==1.0.0
statsmodels==0.14.1 